[FIX] Truncate or pad hexadecimal default-/actualValues.
The attributes "defaultValue" or "actualValue" of an XDD's Object or
SubObject are of type xs:string and can be of arbitrary length. Take
this fact into consideration when processing these values and convert
them to hexadecimal strings of a length that corresponds to an Object's
dataType. I.e.: 0xFF for an UINT8 instead of 0xF or 0x00FF.